requirements.txt
# Framework Web
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-CORS==4.0.0

# Banco de Dados MySQL
mysql-connector-python==8.2.0
PyMySQL==1.1.0

# Validação e Formulários
WTForms==3.1.0
email-validator==2.1.0

# Criptografia e Segurança
bcrypt==4.1.2
cryptography==41.0.7

# Utilitários
python-dotenv==1.0.0
Werkzeug==3.0.1

# Data e Tempo
python-dateutil==2.8.2

# Desenvolvimento e Debug
Flask-DebugToolbar==0.13.1

# Testes (opcional)
pytest==7.4.3
pytest-flask==1.3.0

# Formatação e Linting (desenvolvimento)
black==23.11.0
flake8==6.1.0